fixed calculations.go
