+++ test/test-tables/pool_units.json
@@ -8,6 +8,15 @@
     "P": "existing Pool Units"
   },
   "PoolUnits": [
+
+    {
+      "r": "0",
+      "a": "10000000",
+      "R": "100000000000000000000000000",
+      "A": "1000000000",
+      "P": "100000000000000000000000000",
+      "expected": "495049504950495049500000"
+    },
     {
       "r": "1000000000",
       "a": "10000000",
